backend-python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.10
langchain-community==0.0.25
langchain-openai==0.0.5
pymilvus==2.3.4
sqlalchemy==2.0.25
pymysql==1.1.0
python-dotenv==1.0.0
apscheduler==3.10.4
sse-starlette==1.8.2
